Add files via upload
Given a CFG as an input file in which each line is having exactly one production rule and epsilon as #, Write a c program to compute the first and follow sets of both T & NT (Terminals & Non-terminals) in the given CFG. (NT-Capital letters, T-small letters).

Attaching the Algorithms/pseudo codes for computing first and follow sets as discussed in the class with an example.

Also attaching the example CFG discussed in class. Your program should work for any CFG that can be given as an i/p file.